pyproject.toml
[tool.poetry]
name = "course-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
black = "^24.4.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
ipython = "^8.26.0"
redis = "^5.0.7"
coverage = "^7.6.0"
setuptools = "^71.1.0"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
